+++ src/main/java/de/teragam/jfxshader/internal/DefaultEffectRenderer.java
@@ -4,10 +4,8 @@ import java.util.List;
 
 import com.sun.javafx.geom.Rectangle;
 import com.sun.javafx.geom.transform.BaseTransform;
-import com.sun.scenario.effect.Effect;
 import com.sun.scenario.effect.FilterContext;
 import com.sun.scenario.effect.ImageData;
-import com.sun.scenario.effect.InternalCoreEffectBase;
 import com.sun.scenario.effect.impl.Renderer;
 import com.sun.scenario.effect.impl.state.RenderState;
 
@@ -19,14 +17,12 @@ import de.teragam.jfxshader.ShaderEffectPeer;
 public class DefaultEffectRenderer implements IEffectRenderer {
 
     @Override
-    public ImageData render(Effect effect, FilterContext fctx, BaseTransform transform, Rectangle outputClip, RenderState rstate, ImageData... inputs) {
-        if (effect instanceof InternalCoreEffectBase) {
-            final List<Class<? extends ShaderEffectPeer<?>>> peerList = ShaderController.getPeerDependencies(
-                    ((InternalCoreEffectBase) effect).getEffect().getClass());
-            if (!peerList.isEmpty()) {
-                return Renderer.getRenderer(fctx).getPeerInstance(fctx, peerList.get(0).getAnnotation(EffectPeer.class).value(), -1)
-                        .filter(effect, rstate, transform, outputClip, inputs);
-            }
+    public ImageData render(InternalEffect effect, FilterContext fctx, BaseTransform transform, Rectangle outputClip, RenderState rstate, ImageData... inputs) {
+        final List<Class<? extends ShaderEffectPeer<?>>> peerList = ShaderController.getPeerDependencies(
+                effect.getEffect().getClass());
+        if (!peerList.isEmpty()) {
+            return Renderer.getRenderer(fctx).getPeerInstance(fctx, peerList.get(0).getAnnotation(EffectPeer.class).value(), -1)
+                    .filter(effect, rstate, transform, outputClip, inputs);
         }
         return new ImageData(fctx, null, null);
     }